multibody: Move `rigid_body_tree` and `parsers` here.
