+++ bindings/pydrake/BUILD.bazel
@@ -76,29 +76,15 @@ drake_pybind_library(
     ],
 )
 
-drake_pybind_library(
-    name = "rbtree_py",
-    cc_deps = [
-        ":autodiff_types_pybind",
+drake_py_library(
+    name = "backwards_compatibility_py",
+    srcs = [
+        "parsers.py",
+        "rbtree.py",
     ],
-    cc_srcs = ["rbtree_py.cc"],
-    package_info = PACKAGE_INFO,
-    py_deps = [
-        ":autodiffutils_py",
-        ":common_py",
-        ":parsers_py",
-        "//bindings/pydrake/multibody:shapes_py",
-    ],
-    py_srcs = ["rbtree.py"],
-)
-
-drake_pybind_library(
-    name = "parsers_py",
-    cc_so_name = "parsers",
-    cc_srcs = ["parsers_py.cc"],
-    package_info = PACKAGE_INFO,
-    py_deps = [
-        ":common_py",
+    deps = [
+        "//bindings/pydrake/multibody:parsers_py",
+        "//bindings/pydrake/multibody:rigid_body_tree_py",
     ],
 )
 
@@ -124,8 +110,6 @@ drake_pybind_library(
 PY_LIBRARIES_WITH_INSTALL = adjust_labels_for_drake_hoist([
     ":autodiffutils_py",
     ":common_py",
-    ":parsers_py",
-    ":rbtree_py",
     ":symbolic_py",
     "//drake/bindings/pydrake/examples",
     "//drake/bindings/pydrake/multibody",
@@ -135,7 +119,9 @@ PY_LIBRARIES_WITH_INSTALL = adjust_labels_for_drake_hoist([
     "//drake/bindings/pydrake/util",
 ])
 
-PY_LIBRARIES = []
+PY_LIBRARIES = [
+    ":backwards_compatibility_py",
+]
 
 # Symbol roll-up (for user ease).
 drake_py_library(
@@ -189,15 +175,6 @@ drake_py_test(
 )
 
 drake_py_test(
-    name = "atlas_constructor_test",
-    size = "small",
-    srcs = ["test/atlas_constructor_test.py"],
-    data = ["//drake/examples/atlas:models"],
-    main = "test/atlas_constructor_test.py",
-    deps = [":rbtree_py"],
-)
-
-drake_py_test(
     name = "common_test",
     size = "small",
     srcs = ["test/common_test.py"],
@@ -233,24 +210,6 @@ drake_py_test(
 )
 
 drake_py_test(
-    name = "rbm_forward_kin_test",
-    size = "small",
-    srcs = ["test/rbt_transform_points_test.py"],
-    data = ["//drake/examples/pendulum:models"],
-    main = "test/rbt_transform_points_test.py",
-    deps = [":rbtree_py"],
-)
-
-drake_py_test(
-    name = "rbt_com_test",
-    size = "small",
-    srcs = ["test/rbt_com_test.py"],
-    data = ["//drake/examples/pendulum:models"],
-    main = "test/rbt_com_test.py",
-    deps = [":rbtree_py"],
-)
-
-drake_py_test(
     name = "symbolic_variables_test",
     size = "small",
     srcs = ["test/symbolic_test.py"],
@@ -259,23 +218,9 @@ drake_py_test(
 )
 
 drake_py_test(
-    name = "parsers_test",
-    size = "small",
-    srcs = ["test/parsers_test.py"],
-    data = [
-        "//drake/examples/acrobot:models",
-        "//drake/examples/pr2:models",
-    ],
-    main = "test/parsers_test.py",
-    deps = [":rbtree_py"],
-)
-
-drake_py_test(
-    name = "rbtree_shim_test",
-    srcs = ["test/rbtree_shim_test.py"],
+    name = "backwards_compatability_test",
     deps = [
-        ":common_py",
-        ":rbtree_py",
+        ":backwards_compatibility_py",
     ],
 )
