+++ bindings/pydrake/all.py
@@ -25,8 +25,6 @@ from . import getDrakePath
 from .autodiffutils import *
 from .common import *
 from .forwarddiff import *
-from .parsers import *
-from .rbtree import *
 from .symbolic import *
 
 # Submodules.
