+++ bindings/pydrake/multibody/all.py
@@ -1,2 +1,4 @@
+from .parsers import *
 from .rigid_body_plant import *
+from .rigid_body_tree import *
 from .shapes import *
