+++ bindings/pydrake/multibody/rigid_body_plant_py.cc
@@ -15,7 +15,7 @@ PYBIND11_MODULE(rigid_body_plant, m) {
   using namespace drake::systems;
 
   // Ensure we have bindings for dependencies.
-  py::module::import("pydrake.rbtree");
+  py::module::import("pydrake.multibody.rigid_body_tree");
   py::module::import("pydrake.systems.framework");
 
   using T = double;
