+++ bindings/pydrake/parsers.py
@@ -1,0 +1,9 @@
+from __future__ import absolute_import
+
+# N.B. This is a shim that will be removed soon-ish.
+
+# TODO(eric.cousineau): On 06/02/2018 (after the underactuated course is done),
+# post a warning that this module will be deprecated.
+# On 07/112018 (after 1.5 months), remove the module.
+
+from .multibody.parsers import *
