+++ bindings/pydrake/solvers/test/pr2_ik_test.py
@@ -3,8 +3,16 @@ from __future__ import absolute_import, division, print_function
 import os
 import numpy as np
 import pydrake
-import pydrake.rbtree
+from pydrake.multibody.parsers import PackageMap
+from pydrake.multibody.rigid_body_tree import (
+    AddModelInstanceFromUrdfStringSearchingInRosPackages,
+    FloatingBaseType,
+    RigidBodyFrame,
+    RigidBodyTree,
+)
 from pydrake.solvers import ik
+
+# TODO(eric.cousineau): Use `unittest` (after moving `ik` into `multibody`).
 
 
 def load_robot_from_urdf(urdf_file):
@@ -14,19 +22,19 @@ def load_robot_from_urdf(urdf_file):
     possible to load a robot with a much simpler syntax
     that uses default values, such as:
 
-      robot = pydrake.rbtree.RigidBodyTree(urdf_file)
+      robot = RigidBodyTree(urdf_file)
 
     """
     urdf_string = open(urdf_file).read()
     base_dir = os.path.dirname(urdf_file)
-    package_map = pydrake.rbtree.PackageMap()
+    package_map = PackageMap()
     weld_frame = None
-    floating_base_type = pydrake.rbtree.kRollPitchYaw
+    floating_base_type = FloatingBaseType.kRollPitchYaw
 
     # Load our model from URDF
-    robot = pydrake.rbtree.RigidBodyTree()
+    robot = RigidBodyTree()
 
-    pydrake.rbtree.AddModelInstanceFromUrdfStringSearchingInRosPackages(
+    AddModelInstanceFromUrdfStringSearchingInRosPackages(
         urdf_string,
         package_map,
         base_dir,
@@ -46,11 +54,9 @@ robot = load_robot_from_urdf(urdf_file)
 
 # Add a convenient frame, positioned 0.1m away from the r_gripper_palm_link
 # along that link's x axis
-robot.addFrame(
-    pydrake.rbtree.RigidBodyFrame("r_hand_frame",
-                                  robot.FindBody("r_gripper_palm_link"),
-                                  np.array([0.1, 0, 0]),
-                                  np.array([0., 0, 0])))
+robot.addFrame(RigidBodyFrame(
+    "r_hand_frame", robot.FindBody("r_gripper_palm_link"),
+    np.array([0.1, 0, 0]), np.array([0., 0, 0])))
 
 # Make sure attribute access works on bodies
 assert robot.world().get_name() == "world"
