--- bindings/pydrake/test/atlas_constructor_test.py
@@ -1,23 +1,0 @@
-from __future__ import absolute_import, division, print_function
-import unittest
-import os.path
-
-from pydrake.parsers import PackageMap
-from pydrake import rbtree
-from pydrake import getDrakePath
-
-
-class TestAtlasConstructor(unittest.TestCase):
-    def test_constructor(self):
-        pm = PackageMap()
-        model = os.path.join(getDrakePath(), "examples", "atlas", "urdf",
-                             "atlas_minimal_contact.urdf")
-        pm.PopulateUpstreamToDrake(model)
-        robot = rbtree.RigidBodyTree(
-            model, package_map=pm,
-            floating_base_type=rbtree.FloatingBaseType.kRollPitchYaw)
-        self.assertEqual(robot.get_num_actuators(), 30)
-
-
-if __name__ == '__main__':
-    unittest.main()
