--- bindings/pydrake/test/rbtree_shim_test.py
@@ -1,7 +1,0 @@
-from __future__ import print_function
-
-import pydrake
-
-assert __name__ == '__main__'
-# Test `rbtree` compatibility shim.
-print(pydrake.rbtree)
